backend/requirements.txt
Flask==2.3.3
Flask-Login==0.7.0
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
WTForms==3.0.1
email-validator==2.0.0
Werkzeug==2.3.7
